pyproject.toml
[tool.poetry]
name = "keimo"
version = "0.1.0"
description = ""
authors = ["Tommy He <tommy@tommyhe.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.3"
openai = "^0.27.0"
google-cloud-texttospeech = "^2.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
